refactor: 后端代码审查优化 — 权限重命名、urls.py 瘦身、路由认证补全
- 重命名 is_admin() → is_authenticated_user() 消除语义误导
- urls.py 560→80 行，Admin 自定义逻辑提取到 admin_customization.py
- 新增集中式分页工具 pagination.py
- 补全 evidence_sorting/sales_dispute 路由认证
- 删除未使用的 RequestMetricsMiddleware
- 更新版本号至 26.46.4，更新 CHANGELOG
