+++ CHANGELOG.md
@@ -2,6 +2,46 @@
 
 本项目的所有重要更改都将记录在此文件中。
 
+## [26.46.4] - 2026-05-09
+
+### 前端
+
+#### 优化
+
+- **共享 UI 组件提取**：新增 DropZone、DetailField、DetailCard 共享组件，消除文档识别、客户、合同等模块的重复实现
+- **CRUD mutations 工厂函数**：新增 `createCrudMutations` 泛型工具，精简 client/lawfirm/team/reminder 等模块的 mutations 模板代码
+- **剪贴板操作统一**：提取 `copyToClipboard` 到 `lib/clipboard.ts`，替代分散的 `navigator.clipboard` + `alert` 调用
+- **金额格式化统一**：使用 `formatAmountInt` 替代内联 `toLocaleString`，保持全站一致
+- **API 调用去硬编码**：消除 raw fetch 中的硬编码 URL 和 localStorage 直接访问，统一通过 `lib/api.ts` 封装
+- **工具函数提取**：新增 `lib/date.ts`、`lib/format.ts`、`lib/download.ts`、`lib/file-utils.ts`，集中管理格式化、日期、下载等通用逻辑
+- **工作台 store 拆分**：从 `workbench-store.ts` 提取 `message-factory.ts` 和 `streaming-helpers.ts`，降低单文件复杂度
+- **LPR 计算器拆分**：提取工具函数到 `tools/utils/lpr.ts`，配置提示常量到 `settings/constants/config-hints.ts`
+
+#### 修复
+
+- **MCP 工具调用错误处理**：优化错误消息展示
+- **工作台错误消息**：改进错误消息显示
+
+#### 移除
+
+- **工作台建议提示卡片**：删除 SuggestedPrompts 组件
+
+### 后端
+
+#### 优化
+
+- **权限方法重命名**：`PermissionMixin.is_admin()` 和 `AuthzUserMixin.is_admin()` 重命名为 `is_authenticated_user()`，消除语义误导（原方法对任何已登录用户返回 True，与 `admin_access.is_admin_user()` 的实际管理员检查混淆）
+- **urls.py 瘦身**：从 560 行精简至 80 行，Admin 侧边栏自定义逻辑（排序、虚拟菜单、Hub 页、工具收藏）提取到独立的 `admin_customization.py` 模块
+- **集中式分页工具**：新增 `apps/core/api/pagination.py`，提供 `PageParams`、`PaginatedOut` schema 和 `paginate_queryset` 辅助函数
+
+#### 修复
+
+- **路由认证补全**：为 `evidence_sorting` 和 `sales_dispute` 路由添加 `JWTOrSessionAuth` 认证，修复未认证即可访问的漏洞
+
+#### 移除
+
+- **死代码清理**：删除已定义但从未加入 middleware stack 的 `RequestMetricsMiddleware`
+
 ## [26.46.3] - 2026-05-08
 
 ### 前端
