+++ README.md
@@ -1,4 +1,4 @@
-# 法穿 AI Copilot · V26.46.3
+# 法穿 AI Copilot · V26.46.4
 
 *由一线执业律师主导研发 · 开源 · 私有化部署 · 数据自持*
