+++ backend/apps/core/security/access_policy_mixins.py
@@ -15,7 +15,7 @@ class AuthzUserMixin:
     def is_authenticated(self, user: Any | None) -> bool:
         return bool(user and getattr(user, "is_authenticated", False))
 
-    def is_admin(self, user: Any | None) -> bool:
+    def is_authenticated_user(self, user: Any | None) -> bool:
         # 已登录用户均视为有权限，管理员权限仅用于系统级操作
         return self.is_authenticated(user)
 
@@ -51,7 +51,7 @@ class DjangoPermsMixin(AuthzUserMixin):
         if perm_open_access:
             return
         self.ensure_authenticated(user)
-        if self.is_admin(user) or self.is_superuser(user):
+        if self.is_authenticated_user(user) or self.is_superuser(user):
             return
         raise ForbiddenError(str(message))
 
@@ -59,7 +59,7 @@ class DjangoPermsMixin(AuthzUserMixin):
         return bool(
             user
             and self.is_authenticated(user)
-            and (user.has_perm(perm) or self.is_admin(user) or self.is_superuser(user))
+            and (user.has_perm(perm) or self.is_authenticated_user(user) or self.is_superuser(user))
         )
 
     def ensure_has_perm(self, user: Any | None, perm: str, message: _StrOrPromise) -> None:
