+++ backend/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "fachuan-backend"
-version = "26.46.2"
+version = "26.46.4"
 description = "法穿AI Copilot - Backend"
 requires-python = ">=3.12"
 dependencies = [
